# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  gredis.Asset:
    properties:
      act_long_movie_url:
        description: 长视频 url
        type: string
      act_open_apps:
        description: 需要下载打开的应用  （json）
      act_pop_time:
        description: 二维码自动弹出时间，单位：秒
        type: integer
      act_qrcode_bg_url:
        description: 二维码背景图url
        type: string
      act_qrcode_org_url:
        description: 二维码原链接url
        type: string
      act_qrcode_url:
        description: 二维码地址
        type: string
      act_toast:
        description: OK键引导文案
        type: string
      act_type:
        description: OK键动作类型 1-无动作 2-打开/下载应用 3-弹出二维码 4-加载长视频
        type: integer
      ban_channel_code:
        description: 排除渠道 - json数组
        type: string
      channel_code:
        description: 对应的渠道 - 全部为 ALL - json数组
        type: string
      cover_url:
        description: 封面图地址
        type: string
      del_flag:
        description: 是否删除  0:否  1:是
        type: integer
      display_order:
        description: 排序
        example: 0
        type: integer
      duration_end_date:
        description: 资源有效期 - 结束时间
        type: string
      duration_start_date:
        description: 资源有效期 - 开始时间
        type: string
      id:
        type: integer
      img_stay_time:
        description: 单张图片停留时长(秒)
        type: integer
      last_update_time:
        description: 最后更新时间 - 排序使用
        type: string
      movie_url:
        description: 视频地址
        type: string
      name:
        description: 名称
        type: string
      own_channel_id:
        description: 素材所属渠道
        type: integer
      pic_urls:
        description: 多张图片url  （json）
      priority:
        description: 优先级  0:优先调用  1:优先下载
        type: integer
      remark:
        description: 描述
        type: string
      score:
        description: 评分
      screen_control_count:
        description: 周期内频控次数
        example: 0
        type: integer
      screen_control_cycle:
        description: 频控周期  天数  自然日
        example: 0
        type: integer
      shelf_status:
        description: 上架状态 1-未上架 2-已上架 3-已下架
        type: integer
      type:
        description: 类型 1-视频 2-图片
        type: integer
      view_cities:
        description: 地区 （json）
        type: string
      view_limit:
        description: 青少年观影限制 0  - 不限制 1 – 限制
        type: integer
      view_tags:
        description: 数据标签 （json）
        type: string
    required:
    - act_type
    - duration_end_date
    - duration_start_date
    - id
    - last_update_time
    - shelf_status
    - type
    type: object
  gredis.TrailerListParam:
    properties:
      channel_code:
        description: 渠道码
        type: string
      device_no:
        description: 设备号
        type: string
      isSecure:
        description: 判断返回链接形式 https or http;
        type: boolean
      page:
        description: 页码
        example: 1
        type: integer
      page_size:
        description: 每页数量
        example: 20
        type: integer
      region_code:
        description: region_code
        type: string
    required:
    - channel_code
    - device_no
    - page
    type: object
  model.EData:
    properties:
      ed:
        type: string
      ek:
        type: string
      iv:
        type: string
      sv:
        type: string
    required:
    - ed
    - ek
    - iv
    type: object
  model.EDataResponse:
    properties:
      ed:
        type: string
      ek:
        type: string
      iv:
        type: string
    type: object
  stats_service.AppLog:
    properties:
      channel_code:
        description: 渠道号
        type: string
      device_no:
        description: 设备号
        type: string
      log_type:
        description: log 类型    1:沙发桌面  2:布丁屏保
        type: integer
      os_version_code:
        description: 系统版本
        type: string
      url:
        description: Log url
        type: string
    required:
    - channel_code
    - device_no
    - log_type
    - os_version_code
    type: object
  stats_service.Device:
    properties:
      android_version_code:
        description: Android 版本 Code
        type: string
      android_version_name:
        description: Android 版本名称
        type: string
      app_name:
        description: APP 名称
        type: string
      app_version_code:
        description: APP 版本 Code
        type: string
      app_version_name:
        description: APP 版本名称
        type: string
      channel_code:
        description: 渠道码
        type: string
      cpu_arch:
        description: CpuArch
        type: string
      device_model:
        description: 设备型号
        type: string
      device_name:
        description: 设备名称
        type: string
      device_no:
        description: 设备号
        type: string
      device_vendor:
        description: 设备厂商
        type: string
      ip:
        description: IP
        type: string
      mac:
        description: MAC
        type: string
      resolution:
        description: 分辨率
        type: string
      signature:
        description: 签名
        type: string
    required:
    - android_version_code
    - android_version_name
    - app_name
    - app_version_code
    - app_version_name
    - channel_code
    - device_model
    - device_name
    - device_no
    - device_vendor
    - ip
    - resolution
    type: object
  stats_service.SdkError:
    properties:
      app_name:
        description: APP 名称
        type: string
      app_version_code:
        description: APP 版本 Code
        type: string
      app_version_name:
        description: APP 版本名称
        type: string
      channel_code:
        description: 渠道码
        type: string
      crash_log:
        description: Crash 日志
        type: string
      crash_time:
        description: Crash 时间
        type: string
      device_no:
        description: 设备号
        type: string
      ext:
        description: 自定义数据
        type: string
      sdk_name:
        description: SDK 名称
        type: string
      sdk_version_code:
        description: SDK 版本 Code
        type: string
      sdk_version_name:
        description: SDK 版本名称
        type: string
      signature:
        description: 签名
        type: string
      user_id:
        description: USER ID
        type: integer
    required:
    - app_name
    - app_version_code
    - app_version_name
    - channel_code
    - crash_log
    - crash_time
    - device_no
    - sdk_name
    - sdk_version_code
    - sdk_version_name
    type: object
  stats_service.SdkEvent:
    properties:
      app_name:
        description: APP 名称
        type: string
      app_version_code:
        description: APP 版本 Code
        type: string
      app_version_name:
        description: APP 版本名称
        type: string
      channel_code:
        description: 渠道码
        type: string
      client_time:
        description: 客户端时间 格式：2020-12-12 12:12:12
        type: string
      cpu_arch:
        description: CpuArch
        type: string
      device_brand:
        description: 设备品牌
        type: string
      device_model:
        description: 设备型号
        type: string
      device_no:
        description: 设备号
        type: string
      event_info:
        description: '参数和参数值数据 -json数组 - 格式：[{"event_kv_json":{"key1": "val1", "key2":
          "val2"},"event_name":"sss","event_type":1},...]'
        type: string
      event_kv_json:
        description: '参数和参数值数据 -json数组 - 格式：{"key1": "val1", "key2": "val2"}'
        type: string
      event_name:
        description: 事件名称
        type: string
      imei:
        description: IMEI
        type: string
      ip:
        description: IP
        type: string
      mac:
        description: MAC 地址
        type: string
      net_type:
        description: "网络类型\tWIFI/4G/5G"
        type: string
      newevent_type:
        description: 事件类型  0:自定义事件,1:预置事件
        type: integer
      newpuid:
        description: "IDFA \t\t\t\t\t\tstring `json:\"idfa\" `"
        type: string
      newsession_id:
        description: 会话 ID
        type: string
      os_version_code:
        description: 操作系统版本 Code
        type: string
      os_version_name:
        description: 操作系统版本名称
        type: string
      page_name:
        description: 所在页面
        type: string
      screen_height:
        description: 分辨率 高度
        type: integer
      screen_width:
        description: 分辨率 宽度
        type: integer
      sdk_name:
        description: SDK 名称
        type: string
      sdk_version_code:
        description: SDK 版本 Code
        type: string
      sdk_version_name:
        description: SDK 版本名称
        type: string
      signature:
        description: 签名
        type: string
    required:
    - app_name
    - app_version_code
    - app_version_name
    - channel_code
    - device_model
    - device_no
    - net_type
    - newsession_id
    - os_version_code
    - os_version_name
    - screen_height
    - screen_width
    - sdk_name
    - sdk_version_code
    - sdk_version_name
    type: object
info:
  contact: {}
paths:
  /trailer_api/app/get_new_app:
    post:
      description: 获取最新的 APP 版本信息
      operationId: UPDATE APP
      parameters:
      - description: UPDATE_APP
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/model.EData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: UPDATE APP
      tags:
      - APP
  /trailer_api/app/get_trailer_conf:
    get:
      description: 获取配置信息
      operationId: Trailer Conf
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Trailer Conf
      tags:
      - Config
  /trailer_api/console/remove_asset/{id}:
    get:
      description: 清除 Redis 素材数据
      operationId: RemoveAsset
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Remove Redis Asset
      tags:
      - Asset Console
  /trailer_api/console/reset_asset:
    get:
      description: 清洗 Redis 素材数据
      operationId: ResetAsset
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Reset Redis Asset
      tags:
      - Asset Console
  /trailer_api/stats/record_device:
    post:
      description: 设备信息上报
      operationId: Record Device
      parameters:
      - description: Device
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/stats_service.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 设备信息上报
      tags:
      - Stats
  /trailer_api/stats/record_sdk_err:
    post:
      description: SDK 错误信息上报
      operationId: Record SdkError
      parameters:
      - description: SdkError
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/stats_service.SdkError'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: SDK 错误信息上报
      tags:
      - Stats
  /trailer_api/stats/record_sdk_event:
    post:
      description: SDK 事件统计   参数：事件 json 数组
      operationId: Insert SdkEvent
      parameters:
      - description: Events
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/stats_service.SdkEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: SDK 事件统计
      tags:
      - Stats
  /trailer_api/stats/record_secret_sdk_event:
    post:
      description: SDK 事件统计-加密
      operationId: Insert Secret SdkEvent
      parameters:
      - description: Events
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/model.EData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: SDK 事件统计-加密
      tags:
      - Stats
  /trailer_api/stats/upload_app_log:
    post:
      description: APP 应用日志上报
      operationId: Upload APP LOG
      parameters:
      - description: AppLog
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/stats_service.AppLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: APP 应用日志上报
      tags:
      - Stats
  /trailer_api/test/check_interface:
    get:
      description: 测试接口
      operationId: Test
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Test Interface
      tags:
      - Test
  /trailer_api/test/check_secret_interface:
    post:
      description: 加解密测试
      operationId: CheckSecretInterface
      parameters:
      - description: UPDATE_APP
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/model.EDataResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Test CheckSecretInterface
      tags:
      - Test
  /trailer_api/trailer/get_secret_trailer_list:
    post:
      description: 获取预告片信息 - 加密
      operationId: Get SecretTrailerList
      parameters:
      - description: EDataParam
        in: body
        name: TrailerListParam
        required: true
        schema:
          $ref: '#/definitions/model.EData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get SecretTrailerList
      tags:
      - Trailer
  /trailer_api/trailer/get_trailer_list:
    post:
      description: 获取预告片信息 - 不加密
      operationId: Get TrailerList
      parameters:
      - description: TrailerListParam
        in: body
        name: TrailerListParam
        required: true
        schema:
          $ref: '#/definitions/gredis.TrailerListParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get TrailerList
      tags:
      - Trailer
  /trailer_api/trailer/sync_asset:
    post:
      description: 同步 Asset 素材信息
      operationId: Sync Asset
      parameters:
      - description: SyncAsset
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/gredis.Asset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Sync Asset
      tags:
      - Trailer
swagger: "2.0"